@@ -37,25 +37,23 @@ if __name__ == "__main__":
             "click",
         ],
         extras_require={
-            ':python_version < "3.8"': [
-                "typing-extensions",
-            ],
             "amset": ["amset>=0.4.15", "pydash"],
+            "cclib": ["cclib"],
             "docs": [
-                "sphinx==4.3.1",
-                "numpydoc==1.1.0",
-                "m2r2==0.3.1",
+                "sphinx==4.4.0",
+                "numpydoc==1.2",
+                "m2r2==0.3.2",
                 "mistune==0.8.4",
-                "ipython==7.30.1",
-                "FireWorks==1.9.7",
-                "pydata-sphinx-theme==0.7.2",
-                "autodoc_pydantic==1.5.1",
+                "ipython==8.1.1",
+                "FireWorks==2.0.2",
+                "pydata-sphinx-theme==0.8.0",
+                "autodoc_pydantic==1.6.1",
                 "sphinx_panels==0.6.0",
             ],
             "tests": [
-                "pytest==6.2.5",
+                "pytest==7.0.1",
                 "pytest-cov==3.0.0",
-                "FireWorks==1.9.7",
+                "FireWorks==2.0.2",
                 # "amset==0.4.15",
             ],
             "dev": ["pre-commit>=2.12.1"],
@@ -63,9 +61,9 @@ if __name__ == "__main__":
         },
         classifiers=[
             "Programming Language :: Python :: 3",
-            "Programming Language :: Python :: 3.7",
             "Programming Language :: Python :: 3.8",
             "Programming Language :: Python :: 3.9",
+            "Programming Language :: Python :: 3.10",
             "Development Status :: 5 - Production/Stable",
             "Intended Audience :: Science/Research",
             "Intended Audience :: System Administrators",
@@ -74,7 +72,7 @@ if __name__ == "__main__":
             "Topic :: Other/Nonlisted Topic",
             "Topic :: Scientific/Engineering",
         ],
-        python_requires=">=3.7",
+        python_requires=">=3.8",
         tests_require=["pytest"],
         entry_points={
             "console_scripts": [
