@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from copy import deepcopy
 from dataclasses import dataclass, field
 from pathlib import Path
 
@@ -42,12 +43,15 @@ class DoubleRelaxMaker(Maker):
     ----------
     name : str
         Name of the flows produced by this maker.
-    relax_maker : .BaseVaspMaker
-        Maker to use to generate the relaxations.
+    relax_maker1 : .BaseVaspMaker
+        Maker to use to generate the first relaxation.
+    relax_maker2 : .BaseVaspMaker
+        Maker to use to generate the second relaxation.
     """
 
     name: str = "double relax"
-    relax_maker: BaseVaspMaker = field(default_factory=RelaxMaker)
+    relax_maker1: BaseVaspMaker = field(default_factory=RelaxMaker)
+    relax_maker2: BaseVaspMaker = field(default_factory=RelaxMaker)
 
     def make(self, structure: Structure, prev_vasp_dir: str | Path | None = None):
         """
@@ -65,15 +69,29 @@ class DoubleRelaxMaker(Maker):
         Flow
             A flow containing two relaxations.
         """
-        relax1 = self.relax_maker.make(structure, prev_vasp_dir=prev_vasp_dir)
+        relax1 = self.relax_maker1.make(structure, prev_vasp_dir=prev_vasp_dir)
         relax1.name += " 1"
 
-        relax2 = self.relax_maker.make(
+        relax2 = self.relax_maker2.make(
             relax1.output.structure, prev_vasp_dir=relax1.output.dir_name
         )
         relax2.name += " 2"
 
         return Flow([relax1, relax2], relax2.output, name=self.name)
+
+    @classmethod
+    def from_relax_maker(cls, relax_maker: BaseVaspMaker):
+        """
+        Instantiate the DoubleRelaxMaker with two relax makers of the same type.
+
+        Parameters
+        ----------
+        relax_maker : .BaseVaspMaker
+            Maker to use to generate the first and second relaxations.
+        """
+        return cls(
+            relax_maker1=deepcopy(relax_maker), relax_maker2=deepcopy(relax_maker)
+        )
 
 
 @dataclass
