@@ -24,10 +24,12 @@ class RunType(ValueEnum):
     HSE06 = "HSE06"
     PB0 = "PB0"
     M06L = "M06L"
-    MBJL = "MBJL"
+    MBJ = "MBJ"
     MS0 = "MS0"
     MS1 = "MS1"
     MS2 = "MS2"
+    RSCAN = "RSCAN"
+    R2SCAN = "R2SCAN"
     RTPSS = "RTPSS"
     SCAN = "SCAN"
     TPSS = "TPSS"
@@ -36,8 +38,9 @@ class RunType(ValueEnum):
     optB88_vdW = "optB88-vdW"
     optPBE_vdW = "optPBE-vdW"
     rev_vdW_DF2 = "rev-vdW-DF2"
-    revPBE_vdW = "revPBE-vdW"
+    vdW_DF = "vdW-DF"
     vdW_DF2 = "vdW-DF2"
+    BEEF_vdW = "BEEF-vdW"
     AM05_U = "AM05+U"
     GGA_U = "GGA+U"
     PBE_U = "PBE+U"
@@ -53,10 +56,12 @@ class RunType(ValueEnum):
     HSE06_U = "HSE06+U"
     PB0_U = "PB0+U"
     M06L_U = "M06L+U"
-    MBJL_U = "MBJL+U"
+    MBJ_U = "MBJ+U"
     MS0_U = "MS0+U"
     MS1_U = "MS1+U"
     MS2_U = "MS2+U"
+    RSCAN_U = "RSCAN+U"
+    R2SCAN_U = "R2SCAN+U"
     RTPSS_U = "RTPSS+U"
     SCAN_U = "SCAN+U"
     TPSS_U = "TPSS+U"
@@ -65,8 +70,9 @@ class RunType(ValueEnum):
     optB88_vdW_U = "optB88-vdW+U"
     optPBE_vdW_U = "optPBE-vdW+U"
     rev_vdW_DF2_U = "rev-vdW-DF2+U"
-    revPBE_vdW_U = "revPBE-vdW+U"
+    vdW_DF_U = "vdW-DF+U"
     vdW_DF2_U = "vdW-DF2+U"
+    BEEF_vdW_U = "BEEF-vdW+U"
     LDA = "LDA"
     LDA_U = "LDA+U"
 
@@ -255,17 +261,17 @@ class CalcType(ValueEnum):
     M06L_Structure_Optimization = "M06L Structure Optimization"
     M06L_Deformation = "M06L Deformation"
     M06L_Unrecognized = "M06L Unrecognized"
-    MBJL_NSCF_Line = "MBJL NSCF Line"
-    MBJL_NSCF_Uniform = "MBJL NSCF Uniform"
-    MBJL_Dielectric = "MBJL Dielectric"
-    MBJL_DFPT = "MBJL DFPT"
-    MBJL_DFPT_Dielectric = "MBJL DFPT Dielectric"
-    MBJL_NMR_Nuclear_Shielding = "MBJL NMR Nuclear Shielding"
-    MBJL_NMR_Electric_Field_Gradient = "MBJL NMR Electric Field Gradient"
-    MBJL_Static = "MBJL Static"
-    MBJL_Structure_Optimization = "MBJL Structure Optimization"
-    MBJL_Deformation = "MBJL Deformation"
-    MBJL_Unrecognized = "MBJL Unrecognized"
+    MBJ_NSCF_Line = "MBJ NSCF Line"
+    MBJ_NSCF_Uniform = "MBJ NSCF Uniform"
+    MBJ_Dielectric = "MBJ Dielectric"
+    MBJ_DFPT = "MBJ DFPT"
+    MBJ_DFPT_Dielectric = "MBJ DFPT Dielectric"
+    MBJ_NMR_Nuclear_Shielding = "MBJ NMR Nuclear Shielding"
+    MBJ_NMR_Electric_Field_Gradient = "MBJ NMR Electric Field Gradient"
+    MBJ_Static = "MBJ Static"
+    MBJ_Structure_Optimization = "MBJ Structure Optimization"
+    MBJ_Deformation = "MBJ Deformation"
+    MBJ_Unrecognized = "MBJ Unrecognized"
     MS0_NSCF_Line = "MS0 NSCF Line"
     MS0_NSCF_Uniform = "MS0 NSCF Uniform"
     MS0_Dielectric = "MS0 Dielectric"
@@ -299,6 +305,28 @@ class CalcType(ValueEnum):
     MS2_Structure_Optimization = "MS2 Structure Optimization"
     MS2_Deformation = "MS2 Deformation"
     MS2_Unrecognized = "MS2 Unrecognized"
+    RSCAN_NSCF_Line = "RSCAN NSCF Line"
+    RSCAN_NSCF_Uniform = "RSCAN NSCF Uniform"
+    RSCAN_Dielectric = "RSCAN Dielectric"
+    RSCAN_DFPT = "RSCAN DFPT"
+    RSCAN_DFPT_Dielectric = "RSCAN DFPT Dielectric"
+    RSCAN_NMR_Nuclear_Shielding = "RSCAN NMR Nuclear Shielding"
+    RSCAN_NMR_Electric_Field_Gradient = "RSCAN NMR Electric Field Gradient"
+    RSCAN_Static = "RSCAN Static"
+    RSCAN_Structure_Optimization = "RSCAN Structure Optimization"
+    RSCAN_Deformation = "RSCAN Deformation"
+    RSCAN_Unrecognized = "RSCAN Unrecognized"
+    R2SCAN_NSCF_Line = "R2SCAN NSCF Line"
+    R2SCAN_NSCF_Uniform = "R2SCAN NSCF Uniform"
+    R2SCAN_Dielectric = "R2SCAN Dielectric"
+    R2SCAN_DFPT = "R2SCAN DFPT"
+    R2SCAN_DFPT_Dielectric = "R2SCAN DFPT Dielectric"
+    R2SCAN_NMR_Nuclear_Shielding = "R2SCAN NMR Nuclear Shielding"
+    R2SCAN_NMR_Electric_Field_Gradient = "R2SCAN NMR Electric Field Gradient"
+    R2SCAN_Static = "R2SCAN Static"
+    R2SCAN_Structure_Optimization = "R2SCAN Structure Optimization"
+    R2SCAN_Deformation = "R2SCAN Deformation"
+    R2SCAN_Unrecognized = "R2SCAN Unrecognized"
     RTPSS_NSCF_Line = "RTPSS NSCF Line"
     RTPSS_NSCF_Uniform = "RTPSS NSCF Uniform"
     RTPSS_Dielectric = "RTPSS Dielectric"
@@ -387,17 +415,17 @@ class CalcType(ValueEnum):
     rev_vdW_DF2_Structure_Optimization = "rev-vdW-DF2 Structure Optimization"
     rev_vdW_DF2_Deformation = "rev-vdW-DF2 Deformation"
     rev_vdW_DF2_Unrecognized = "rev-vdW-DF2 Unrecognized"
-    revPBE_vdW_NSCF_Line = "revPBE-vdW NSCF Line"
-    revPBE_vdW_NSCF_Uniform = "revPBE-vdW NSCF Uniform"
-    revPBE_vdW_Dielectric = "revPBE-vdW Dielectric"
-    revPBE_vdW_DFPT = "revPBE-vdW DFPT"
-    revPBE_vdW_DFPT_Dielectric = "revPBE-vdW DFPT Dielectric"
-    revPBE_vdW_NMR_Nuclear_Shielding = "revPBE-vdW NMR Nuclear Shielding"
-    revPBE_vdW_NMR_Electric_Field_Gradient = "revPBE-vdW NMR Electric Field Gradient"
-    revPBE_vdW_Static = "revPBE-vdW Static"
-    revPBE_vdW_Structure_Optimization = "revPBE-vdW Structure Optimization"
-    revPBE_vdW_Deformation = "revPBE-vdW Deformation"
-    revPBE_vdW_Unrecognized = "revPBE-vdW Unrecognized"
+    vdW_DF_NSCF_Line = "vdW-DF NSCF Line"
+    vdW_DF_NSCF_Uniform = "vdW-DF NSCF Uniform"
+    vdW_DF_Dielectric = "vdW-DF Dielectric"
+    vdW_DF_DFPT = "vdW-DF DFPT"
+    vdW_DF_DFPT_Dielectric = "vdW-DF DFPT Dielectric"
+    vdW_DF_NMR_Nuclear_Shielding = "vdW-DF NMR Nuclear Shielding"
+    vdW_DF_NMR_Electric_Field_Gradient = "vdW-DF NMR Electric Field Gradient"
+    vdW_DF_Static = "vdW-DF Static"
+    vdW_DF_Structure_Optimization = "vdW-DF Structure Optimization"
+    vdW_DF_Deformation = "vdW-DF Deformation"
+    vdW_DF_Unrecognized = "vdW-DF Unrecognized"
     vdW_DF2_NSCF_Line = "vdW-DF2 NSCF Line"
     vdW_DF2_NSCF_Uniform = "vdW-DF2 NSCF Uniform"
     vdW_DF2_Dielectric = "vdW-DF2 Dielectric"
@@ -409,6 +437,17 @@ class CalcType(ValueEnum):
     vdW_DF2_Structure_Optimization = "vdW-DF2 Structure Optimization"
     vdW_DF2_Deformation = "vdW-DF2 Deformation"
     vdW_DF2_Unrecognized = "vdW-DF2 Unrecognized"
+    BEEF_vdW_NSCF_Line = "BEEF-vdW NSCF Line"
+    BEEF_vdW_NSCF_Uniform = "BEEF-vdW NSCF Uniform"
+    BEEF_vdW_Dielectric = "BEEF-vdW Dielectric"
+    BEEF_vdW_DFPT = "BEEF-vdW DFPT"
+    BEEF_vdW_DFPT_Dielectric = "BEEF-vdW DFPT Dielectric"
+    BEEF_vdW_NMR_Nuclear_Shielding = "BEEF-vdW NMR Nuclear Shielding"
+    BEEF_vdW_NMR_Electric_Field_Gradient = "BEEF-vdW NMR Electric Field Gradient"
+    BEEF_vdW_Static = "BEEF-vdW Static"
+    BEEF_vdW_Structure_Optimization = "BEEF-vdW Structure Optimization"
+    BEEF_vdW_Deformation = "BEEF-vdW Deformation"
+    BEEF_vdW_Unrecognized = "BEEF-vdW Unrecognized"
     AM05_U_NSCF_Line = "AM05+U NSCF Line"
     AM05_U_NSCF_Uniform = "AM05+U NSCF Uniform"
     AM05_U_Dielectric = "AM05+U Dielectric"
@@ -576,17 +615,17 @@ class CalcType(ValueEnum):
     M06L_U_Structure_Optimization = "M06L+U Structure Optimization"
     M06L_U_Deformation = "M06L+U Deformation"
     M06L_U_Unrecognized = "M06L+U Unrecognized"
-    MBJL_U_NSCF_Line = "MBJL+U NSCF Line"
-    MBJL_U_NSCF_Uniform = "MBJL+U NSCF Uniform"
-    MBJL_U_Dielectric = "MBJL+U Dielectric"
-    MBJL_U_DFPT = "MBJL+U DFPT"
-    MBJL_U_DFPT_Dielectric = "MBJL+U DFPT Dielectric"
-    MBJL_U_NMR_Nuclear_Shielding = "MBJL+U NMR Nuclear Shielding"
-    MBJL_U_NMR_Electric_Field_Gradient = "MBJL+U NMR Electric Field Gradient"
-    MBJL_U_Static = "MBJL+U Static"
-    MBJL_U_Structure_Optimization = "MBJL+U Structure Optimization"
-    MBJL_U_Deformation = "MBJL+U Deformation"
-    MBJL_U_Unrecognized = "MBJL+U Unrecognized"
+    MBJ_U_NSCF_Line = "MBJ+U NSCF Line"
+    MBJ_U_NSCF_Uniform = "MBJ+U NSCF Uniform"
+    MBJ_U_Dielectric = "MBJ+U Dielectric"
+    MBJ_U_DFPT = "MBJ+U DFPT"
+    MBJ_U_DFPT_Dielectric = "MBJ+U DFPT Dielectric"
+    MBJ_U_NMR_Nuclear_Shielding = "MBJ+U NMR Nuclear Shielding"
+    MBJ_U_NMR_Electric_Field_Gradient = "MBJ+U NMR Electric Field Gradient"
+    MBJ_U_Static = "MBJ+U Static"
+    MBJ_U_Structure_Optimization = "MBJ+U Structure Optimization"
+    MBJ_U_Deformation = "MBJ+U Deformation"
+    MBJ_U_Unrecognized = "MBJ+U Unrecognized"
     MS0_U_NSCF_Line = "MS0+U NSCF Line"
     MS0_U_NSCF_Uniform = "MS0+U NSCF Uniform"
     MS0_U_Dielectric = "MS0+U Dielectric"
@@ -620,6 +659,28 @@ class CalcType(ValueEnum):
     MS2_U_Structure_Optimization = "MS2+U Structure Optimization"
     MS2_U_Deformation = "MS2+U Deformation"
     MS2_U_Unrecognized = "MS2+U Unrecognized"
+    RSCAN_U_NSCF_Line = "RSCAN+U NSCF Line"
+    RSCAN_U_NSCF_Uniform = "RSCAN+U NSCF Uniform"
+    RSCAN_U_Dielectric = "RSCAN+U Dielectric"
+    RSCAN_U_DFPT = "RSCAN+U DFPT"
+    RSCAN_U_DFPT_Dielectric = "RSCAN+U DFPT Dielectric"
+    RSCAN_U_NMR_Nuclear_Shielding = "RSCAN+U NMR Nuclear Shielding"
+    RSCAN_U_NMR_Electric_Field_Gradient = "RSCAN+U NMR Electric Field Gradient"
+    RSCAN_U_Static = "RSCAN+U Static"
+    RSCAN_U_Structure_Optimization = "RSCAN+U Structure Optimization"
+    RSCAN_U_Deformation = "RSCAN+U Deformation"
+    RSCAN_U_Unrecognized = "RSCAN+U Unrecognized"
+    R2SCAN_U_NSCF_Line = "R2SCAN+U NSCF Line"
+    R2SCAN_U_NSCF_Uniform = "R2SCAN+U NSCF Uniform"
+    R2SCAN_U_Dielectric = "R2SCAN+U Dielectric"
+    R2SCAN_U_DFPT = "R2SCAN+U DFPT"
+    R2SCAN_U_DFPT_Dielectric = "R2SCAN+U DFPT Dielectric"
+    R2SCAN_U_NMR_Nuclear_Shielding = "R2SCAN+U NMR Nuclear Shielding"
+    R2SCAN_U_NMR_Electric_Field_Gradient = "R2SCAN+U NMR Electric Field Gradient"
+    R2SCAN_U_Static = "R2SCAN+U Static"
+    R2SCAN_U_Structure_Optimization = "R2SCAN+U Structure Optimization"
+    R2SCAN_U_Deformation = "R2SCAN+U Deformation"
+    R2SCAN_U_Unrecognized = "R2SCAN+U Unrecognized"
     RTPSS_U_NSCF_Line = "RTPSS+U NSCF Line"
     RTPSS_U_NSCF_Uniform = "RTPSS+U NSCF Uniform"
     RTPSS_U_Dielectric = "RTPSS+U Dielectric"
@@ -718,19 +779,17 @@ class CalcType(ValueEnum):
     rev_vdW_DF2_U_Structure_Optimization = "rev-vdW-DF2+U Structure Optimization"
     rev_vdW_DF2_U_Deformation = "rev-vdW-DF2+U Deformation"
     rev_vdW_DF2_U_Unrecognized = "rev-vdW-DF2+U Unrecognized"
-    revPBE_vdW_U_NSCF_Line = "revPBE-vdW+U NSCF Line"
-    revPBE_vdW_U_NSCF_Uniform = "revPBE-vdW+U NSCF Uniform"
-    revPBE_vdW_U_Dielectric = "revPBE-vdW+U Dielectric"
-    revPBE_vdW_U_DFPT = "revPBE-vdW+U DFPT"
-    revPBE_vdW_U_DFPT_Dielectric = "revPBE-vdW+U DFPT Dielectric"
-    revPBE_vdW_U_NMR_Nuclear_Shielding = "revPBE-vdW+U NMR Nuclear Shielding"
-    revPBE_vdW_U_NMR_Electric_Field_Gradient = (
-        "revPBE-vdW+U NMR Electric Field Gradient"
-    )
-    revPBE_vdW_U_Static = "revPBE-vdW+U Static"
-    revPBE_vdW_U_Structure_Optimization = "revPBE-vdW+U Structure Optimization"
-    revPBE_vdW_U_Deformation = "revPBE-vdW+U Deformation"
-    revPBE_vdW_U_Unrecognized = "revPBE-vdW+U Unrecognized"
+    vdW_DF_U_NSCF_Line = "vdW-DF+U NSCF Line"
+    vdW_DF_U_NSCF_Uniform = "vdW-DF+U NSCF Uniform"
+    vdW_DF_U_Dielectric = "vdW-DF+U Dielectric"
+    vdW_DF_U_DFPT = "vdW-DF+U DFPT"
+    vdW_DF_U_DFPT_Dielectric = "vdW-DF+U DFPT Dielectric"
+    vdW_DF_U_NMR_Nuclear_Shielding = "vdW-DF+U NMR Nuclear Shielding"
+    vdW_DF_U_NMR_Electric_Field_Gradient = "vdW-DF+U NMR Electric Field Gradient"
+    vdW_DF_U_Static = "vdW-DF+U Static"
+    vdW_DF_U_Structure_Optimization = "vdW-DF+U Structure Optimization"
+    vdW_DF_U_Deformation = "vdW-DF+U Deformation"
+    vdW_DF_U_Unrecognized = "vdW-DF+U Unrecognized"
     vdW_DF2_U_NSCF_Line = "vdW-DF2+U NSCF Line"
     vdW_DF2_U_NSCF_Uniform = "vdW-DF2+U NSCF Uniform"
     vdW_DF2_U_Dielectric = "vdW-DF2+U Dielectric"
@@ -742,6 +801,17 @@ class CalcType(ValueEnum):
     vdW_DF2_U_Structure_Optimization = "vdW-DF2+U Structure Optimization"
     vdW_DF2_U_Deformation = "vdW-DF2+U Deformation"
     vdW_DF2_U_Unrecognized = "vdW-DF2+U Unrecognized"
+    BEEF_vdW_U_NSCF_Line = "BEEF-vdW+U NSCF Line"
+    BEEF_vdW_U_NSCF_Uniform = "BEEF-vdW+U NSCF Uniform"
+    BEEF_vdW_U_Dielectric = "BEEF-vdW+U Dielectric"
+    BEEF_vdW_U_DFPT = "BEEF-vdW+U DFPT"
+    BEEF_vdW_U_DFPT_Dielectric = "BEEF-vdW+U DFPT Dielectric"
+    BEEF_vdW_U_NMR_Nuclear_Shielding = "BEEF-vdW+U NMR Nuclear Shielding"
+    BEEF_vdW_U_NMR_Electric_Field_Gradient = "BEEF-vdW+U NMR Electric Field Gradient"
+    BEEF_vdW_U_Static = "BEEF-vdW+U Static"
+    BEEF_vdW_U_Structure_Optimization = "BEEF-vdW+U Structure Optimization"
+    BEEF_vdW_U_Deformation = "BEEF-vdW+U Deformation"
+    BEEF_vdW_U_Unrecognized = "BEEF-vdW+U Unrecognized"
     LDA_NSCF_Line = "LDA NSCF Line"
     LDA_NSCF_Uniform = "LDA NSCF Uniform"
     LDA_Dielectric = "LDA Dielectric"
